refactoring todo app, moving to sub folders. changning components to functions
