+++ src/containers/Todo/todos.constants.js
@@ -1,0 +1,3 @@
+
+export const ADD_TODO = 'Todo/ADD_TODO';
+export const REMOVE_TODO = 'Todo/REMOVE_TODO';
