+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux'
-import rootReducer from './reducers'
+import rootReducer from './rootReducer'
 import './index.css';
 import App from './components/App';
